==> jupyter_notebook/1-GeneSys_prepare_inputs_based_on_cell_lineage_blueprint.ipynb <==
import os

## Set directory
os.chdir('/hpc/group/pbenfeylab/CheWei/CW_data/genesys')

from genesys_train_v1 import *
%matplotlib inline
%config InlineBackend.figure_format = 'retina'

## Conda Env pytorch-gpu on DCC
## pytorch version
print(torch.__version__)
## scanpy
print(sc.__version__) 

## Genes considered/used (shared among samples) 
gene_list = pd.read_csv('./gene_list_1108.csv')

#### Load root atlas data that is going to used for GeneSys training

adata = sc.read_h5ad("/hpc/group/pbenfeylab/CheWei/scRNA-seq/Integrated_Objects/rc.integrated.integrated.h5ad")
meta = pd.read_csv("/hpc/group/pbenfeylab/CheWei/scRNA-seq/Integrated_Objects/Root_Atlas_meta.csv", index_col=0)
adata.obs['celltype.anno'] = meta['celltype.anno']
adata.obs['consensus.time'] = meta['pseudotime']

adata

##The integrated assay values were floored at zero and normalized to the range of zero to one
adata.X[adata.X < 0]=0
mmin = np.amin(adata.X)
nor = (np.amax(adata.X)-mmin)
adata.X = (adata.X-mmin)/nor

## Prepare labels
time_bin = adata.obs['consensus.time'].copy()
celltype = adata.obs['celltype.anno']
celltype = celltype.replace('Quiescent Center', -1)
celltype = celltype.replace('Columella', 0)
celltype = celltype.replace('Lateral Root Cap', 1)
celltype = celltype.replace('Phloem', 2)
celltype = celltype.replace('Xylem', 3)
celltype = celltype.replace('Procambium', 4)
celltype = celltype.replace('Pericycle', 5)
celltype = celltype.replace('Endodermis', 6)
celltype = celltype.replace('Cortex', 7)
celltype = celltype.replace('Atrichoblast', 8)
celltype = celltype.replace('Trichoblast', 9)
for i in range(10):
    time_bin[np.where(celltype == i)[0]]=pd.qcut(pd.Series(adata.obs['consensus.time'][np.where(celltype == i)[0]]),q=10,labels=range(1,11)).to_numpy()
    
time_bin[np.where(celltype == -1)[0]] = 0 
labels = adata.obs['celltype.anno'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)

## Prepare training, validation and test set
X_train, X_test, y_train, y_test = train_test_split(
            np.array(adata.X), labels, test_size=0.2, stratify=labels, shuffle=True, random_state=0)
X_train, X_val, y_train, y_val = train_test_split(
            X_train, y_train, test_size=0.2, stratify=y_train, shuffle=True, random_state=0)
data = {}
data['X_train'] = X_train
data['y_train'] = y_train
data['X_val'] = X_val
data['y_val'] = y_val
data['X_test'] = X_test
data['y_test'] = y_test

## Save the input data for GeneSys

#with open("./genesys_root_data.pkl", 'wb') as file_handle:
#    pickle.dump(data, file_handle)

#### Load unseen scRNA-seq data

## Load unseen WT scRNA-seq data (The .h5ad file is converted from Seurat Objects from GEO GSE152766)
wt = sc.read_h5ad("/hpc/group/pbenfeylab/CheWei/scRNA-seq/Integrated_Objects/rswt_integrated.h5ad")

wt

##The integrated assay values were floored at zero and normalized to the range of zero to one
wt.X[wt.X < 0]=0
mmin = np.amin(wt.X)
nor = (np.amax(wt.X)-mmin)
wt.X = (wt.X-mmin)/nor

## Resize the number of features/genes to match GeneSys's input dimension for features/genes
mi = match(np.array(gene_list['features']).tolist(),wt.var.features.tolist())
mi = np.array(mi)

wtx = pd.DataFrame(wt.X.copy())
wtx[len(wtx.columns)] = pd.Series(0.0, index=np.arange(len(wtx)))
mi[np.where(mi==None)[0]] = len(wtx.columns)-1
wtx = wtx[pd.Series(mi)]
wtx.columns = np.array(gene_list['features']).tolist()

wtx.shape

wtx

## Prepare labels
time_bin = wt.obs['CytoTRACE.time'].copy()
celltype = wt.obs['celltype.anno.crude']
celltype = celltype.replace('Quiescent Center', -1)
celltype = celltype.replace('Columella', 0)
celltype = celltype.replace('Lateral Root Cap', 1)
celltype = celltype.replace('Phloem', 2)
celltype = celltype.replace('Xylem', 3)
celltype = celltype.replace('Procambium', 4)
celltype = celltype.replace('Pericycle', 5)
celltype = celltype.replace('Endodermis', 6)
celltype = celltype.replace('Cortex', 7)
celltype = celltype.replace('Atrichoblast', 8)
celltype = celltype.replace('Trichoblast', 9)
for i in range(10):
    time_bin[np.where(celltype == i)[0]]=pd.qcut(pd.Series(wt.obs['CytoTRACE.time'][np.where(celltype == i)[0]]),q=10,labels=range(1,11)).to_numpy()
    
time_bin[np.where(celltype == -1)[0]] = 0 
labels = wt.obs['celltype.anno.crude'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)

## Prepare training, validation and test set
X_train, X_test, y_train, y_test = train_test_split(
            np.array(wtx), labels, test_size=0.2, stratify=labels, shuffle=True, random_state=0)
X_train, X_val, y_train, y_val = train_test_split(
            X_train, y_train, test_size=0.2, stratify=y_train, shuffle=True, random_state=0)
data = {}
data['X_train'] = X_train
data['y_train'] = y_train
data['X_val'] = X_val
data['y_val'] = y_val
data['X_test'] = X_test
data['y_test'] = y_test

## Save the input data for GeneSys

#with open("./genesys_rswt_data.pkl", 'wb') as file_handle:
#    pickle.dump(data, file_handle)

#### Load mutant scRNA-seq data (shr and scr)

## Load mutant scRNA-seq data (The .h5ad file is converted from Seurat Objects from GEO GSE152766)
shr = sc.read_h5ad("/hpc/group/pbenfeylab/CheWei/scRNA-seq/Integrated_Objects/shr_integrated.h5ad")
scr = sc.read_h5ad("/hpc/group/pbenfeylab/CheWei/scRNA-seq/Integrated_Objects/scr_integrated.h5ad")

shr

scr

# Floored at 0 (And make sure ceiling at 10) and normalize to range 0-1
shr.X[shr.X < 0]=0
mmin = np.amin(shr.X)
nor = (np.amax(shr.X)-mmin)
shr.X = (shr.X-mmin)/nor
############################
scr.X[scr.X < 0]=0
mmin = np.amin(scr.X)
nor = (np.amax(scr.X)-mmin)
scr.X = (scr.X-mmin)/nor

## Resize the number of features/genes to match GeneSys's input dimension for features/genes
mi = match(np.array(gene_list['features']).tolist(),shr.var.features.tolist())
mi = np.array(mi)
shrx = pd.DataFrame(shr.X.copy())
shrx[len(shrx.columns)] = pd.Series(0.0, index=np.arange(len(shrx)))
mi[np.where(mi==None)[0]] = len(shrx.columns)-1
shrx = shrx[pd.Series(mi)]
shrx.columns = np.array(gene_list['features']).tolist()

shrx.shape

## Prepare labels
time_bin = shr.obs['CytoTRACE.time'].copy()
celltype = shr.obs['celltype.anno.crude']
celltype = celltype.replace('Quiescent Center', -1)
celltype = celltype.replace('Columella', 0)
celltype = celltype.replace('Lateral Root Cap', 1)
celltype = celltype.replace('Phloem', 2)
celltype = celltype.replace('Xylem', 3)
celltype = celltype.replace('Procambium', 4)
celltype = celltype.replace('Pericycle', 5)
celltype = celltype.replace('Endodermis', 6)
celltype = celltype.replace('Cortex', 7)
celltype = celltype.replace('Atrichoblast', 8)
celltype = celltype.replace('Trichoblast', 9)
for i in range(10):
    time_bin[np.where(celltype == i)[0]]=pd.qcut(pd.Series(shr.obs['CytoTRACE.time'][np.where(celltype == i)[0]]),q=10,labels=range(1,11)).to_numpy()
    
time_bin[np.where(celltype == -1)[0]] = 0 
labels = shr.obs['celltype.anno.crude'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)

## Prepare training, validation and test set
X_train, X_test, y_train, y_test = train_test_split(
            np.array(shrx), labels, test_size=0.2, stratify=labels, shuffle=True, random_state=0)
X_train, X_val, y_train, y_val = train_test_split(
            X_train, y_train, test_size=0.2, stratify=y_train, shuffle=True, random_state=0)
data = {}
data['X_train'] = X_train
data['y_train'] = y_train
data['X_val'] = X_val
data['y_val'] = y_val
data['X_test'] = X_test
data['y_test'] = y_test

#with open("./genesys_shr_data.pkl", 'wb') as file_handle:
#    pickle.dump(data, file_handle)

## Resize the number of features/genes to match GeneSys's input dimension for features/genes
mi = match(np.array(gene_list['features']).tolist(),scr.var.features.tolist())
mi = np.array(mi)

scrx = pd.DataFrame(scr.X.copy())
scrx[len(scrx.columns)] = pd.Series(0.0, index=np.arange(len(scrx)))
mi[np.where(mi==None)[0]] = len(scrx.columns)-1
scrx = scrx[pd.Series(mi)]
scrx.columns = np.array(gene_list['features']).tolist()

scrx.shape

## Prepare labels
time_bin = scr.obs['CytoTRACE.time'].copy()
celltype = scr.obs['celltype.anno.crude']
#celltype = celltype.replace('Quiescent Center', -1)
celltype = celltype.replace('Quiescent Center', 0)
celltype = celltype.replace('Columella', 0)
celltype = celltype.replace('Lateral Root Cap', 1)
celltype = celltype.replace('Phloem', 2)
celltype = celltype.replace('Xylem', 3)
celltype = celltype.replace('Procambium', 4)
celltype = celltype.replace('Pericycle', 5)
celltype = celltype.replace('Endodermis', 6)
celltype = celltype.replace('Cortex', 7)
celltype = celltype.replace('Atrichoblast', 8)
celltype = celltype.replace('Trichoblast', 9)
for i in range(10):
    time_bin[np.where(celltype == i)[0]]=pd.qcut(pd.Series(scr.obs['CytoTRACE.time'][np.where(celltype == i)[0]]),q=10,labels=range(1,11)).to_numpy()
    
time_bin[np.where(celltype == -1)[0]] = 0 
labels = shr.obs['celltype.anno.crude'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)

## Prepare training, validation and test set
X_train, X_test, y_train, y_test = train_test_split(
            np.array(scrx), labels, test_size=0.2, stratify=labels, shuffle=True, random_state=0)
X_train, X_val, y_train, y_val = train_test_split(
            X_train, y_train, test_size=0.2, stratify=y_train, shuffle=True, random_state=0)
data = {}
data['X_train'] = X_train
data['y_train'] = y_train
data['X_val'] = X_val
data['y_val'] = y_val
data['X_test'] = X_test
data['y_test'] = y_test

#with open("./genesys_scr_data.pkl", 'wb') as file_handle:
#    pickle.dump(data, file_handle)
