==> jupyter_notebook/5-Prepare_GeneSys_Input_Mouse_Embryo_Data_based_on_cell_lineage_blueprint_2500_GPC_UPC_20240301.ipynb <==
import os

## Set directory
os.chdir('/hpc/group/pbenfeylab/CheWei/CW_data/genesys')

from genesys import *
%matplotlib inline
%config InlineBackend.figure_format = 'retina'

#import scvelo as scv
#import cellrank as cr
import seaborn as sns
from sklearn.utils import resample

## Conda Env pytorch-gpu on DCC
## pytorch version
print(torch.__version__)
## scanpy
print(sc.__version__)
## scvelo
#print(scv.__version__)
## cellrank
#print(cr.__version__)

## Cell Lineage Blueprint
#### White Blood Cells include HSCs, treated as precursor for definitive erythroid, B cells, T cells and Mast cells (if precursors for white bllod cells do not exist, set the transcriptomic profiles to zero)
#### Neuroectoderm and glia treated as precursor for CNS neurons,intermediate neuronal progenitors, eye and other, olfactory sensory neurons, oligodendrocytes, and ependymal_cells
#### mesoderm treated as precursor for Adipocytes and Muscule cells
#### Epithelial cells treated as precursor for Lung and airway
#### We remove transient cell population Primitive Erythroid and rarer cell type Testis and Adrenal from the training
#### For state that doesn't have cells captured, we sample cells equally from earier state and later state

## Load meta data

## Genes used 
df_gene = pd.read_csv('./Mouse_Embryo/df_gene.csv')

df_gene

## Cells involved 
df_cell = pd.read_csv('./Mouse_Embryo/df_cell.csv')

df_cell

pd.Series(df_cell['day']).value_counts()

pd.Series(df_cell['major_trajectory']).value_counts()

## Load gene exp matrices

## adata1

adata1 = sc.read_h5ad("./Mouse_Embryo/adata_JAX_dataset_1.h5ad")

adata1

sc.pp.filter_cells(adata1, min_counts=2500)
sc.pp.filter_cells(adata1, min_genes=2500)
# Calculate row sums for the number of genes per cell
row_sums = np.sum(adata1.X > 0 , axis=1).reshape(-1).tolist()[0]
# Calculate row sums for the number of counts per cell
row_sums_counts = np.sum(adata1.X, axis=1).reshape(-1).tolist()[0]

## Median genes per cell
np.median(row_sums)

## Median UMI counts per cell
np.median(row_sums_counts)

adata1

pd.Series(adata1.obs['day']).value_counts()

celltype = pd.merge(adata1.obs['cell_id'], df_cell, on='cell_id', how='left')
celltype.index = adata1.obs.index
adata1.obs['celltype.anno']=celltype['major_trajectory']

## Prepare labels
## high resolution: 11 clusters
#E8.5-9, E9.25-10, E10.25-10.5, E10.75-11.75, E12-12.75, E13-13.75, E14-14.333, E14.75-15.25, E15.5-16.5, E16.75-18, E18.25-P0
time_bin = adata1.obs['day'].copy()
time_bin = time_bin.replace('E8.5', 0)
time_bin = time_bin.replace('E8.75', 0)
time_bin = time_bin.replace('E9.0', 0)
time_bin = time_bin.replace('E9.25', 1)
time_bin = time_bin.replace('E9.5', 1)
time_bin = time_bin.replace('E9.75', 1)
time_bin = time_bin.replace('E10.0', 1)
time_bin = time_bin.replace('E10.25', 2)
time_bin = time_bin.replace('E10.5', 2)
time_bin = time_bin.replace('E10.75', 3)
time_bin = time_bin.replace('E11.0', 3)
time_bin = time_bin.replace('E11.25', 3)
time_bin = time_bin.replace('E11.5', 3)
time_bin = time_bin.replace('E11.75', 3)
time_bin = time_bin.replace('E12.0', 4)
time_bin = time_bin.replace('E12.25', 4)
time_bin = time_bin.replace('E12.5', 4)
time_bin = time_bin.replace('E12.75', 4)
time_bin = time_bin.replace('E13.0', 5)
time_bin = time_bin.replace('E13.25', 5)
time_bin = time_bin.replace('E13.5', 5)
time_bin = time_bin.replace('E13.75', 5)
time_bin = time_bin.replace('E14.0', 6)
time_bin = time_bin.replace('E14.25', 6)
time_bin = time_bin.replace('E14.333', 6)
time_bin = time_bin.replace('E14.75', 7)
time_bin = time_bin.replace('E15.0', 7)
time_bin = time_bin.replace('E15.25', 7)
time_bin = time_bin.replace('E15.5', 8)
time_bin = time_bin.replace('E15.75', 8)
time_bin = time_bin.replace('E16.0', 8)
time_bin = time_bin.replace('E16.25', 8)
time_bin = time_bin.replace('E16.5', 8)
time_bin = time_bin.replace('E16.75', 9)
time_bin = time_bin.replace('E17.0', 9)
time_bin = time_bin.replace('E17.25', 9)
time_bin = time_bin.replace('E17.5', 9)
time_bin = time_bin.replace('E17.75', 9)
time_bin = time_bin.replace('E18.0', 9)
time_bin = time_bin.replace('E18.25', 10)
time_bin = time_bin.replace('E18.5', 10)
time_bin = time_bin.replace('E18.75', 10)
time_bin = time_bin.replace('P0', 10)

#celltype = adata1.obs['celltype.anno']
#celltype = celltype.replace('Mesoderm', 0)
#celltype = celltype.replace('CNS_neurons', 1)
#celltype = celltype.replace('Neuroectoderm_and_glia', 2)
#celltype = celltype.replace('Epithelial_cells', 3)
#celltype = celltype.replace('Muscle_cells', 4)
#celltype = celltype.replace('Endothelium', 5)
#celltype = celltype.replace('Hepatocytes', 6)
#celltype = celltype.replace('Neural_crest_PNS_glia', 7)
#celltype = celltype.replace('Neural_crest_PNS_neurons', 8)
#celltype = celltype.replace('Cardiomyocytes', 9)
#celltype = celltype.replace('Megakaryocytes', 10)
#celltype = celltype.replace('Ependymal_cells', 11)

celltype = adata1.obs['celltype.anno']
celltype = celltype.replace('Mesoderm', 0)
celltype = celltype.replace('CNS_neurons', 1)
celltype = celltype.replace('Neuroectoderm_and_glia', 2)
celltype = celltype.replace('Definitive_erythroid', 3)
celltype = celltype.replace('Epithelial_cells', 4)
celltype = celltype.replace('Intermediate_neuronal_progenitors', 5)
celltype = celltype.replace('Muscle_cells', 6)
celltype = celltype.replace('Endothelium', 7)
celltype = celltype.replace('White_blood_cells', 8)
celltype = celltype.replace('Hepatocytes', 9)
celltype = celltype.replace('Neural_crest_PNS_glia', 10)
celltype = celltype.replace('Neural_crest_PNS_neurons', 11)
celltype = celltype.replace('Adipocytes', 12)
celltype = celltype.replace('Eye_and_other', 13)
celltype = celltype.replace('Lung_and_airway', 14)
celltype = celltype.replace('Intestine', 15)
celltype = celltype.replace('T_cells', 16)
celltype = celltype.replace('Cardiomyocytes', 17)
celltype = celltype.replace('Olfactory_sensory_neurons', 18)
celltype = celltype.replace('B_cells', 19)
celltype = celltype.replace('Mast_cells', 20)
celltype = celltype.replace('Oligodendrocytes', 21)
celltype = celltype.replace('Megakaryocytes', 22)
celltype = celltype.replace('Ependymal_cells', 23)
celltype = celltype.replace('Primitive_erythroid', 24)
celltype = celltype.replace('Testis_and_adrenal', 25) 
 
labels = adata1.obs['celltype.anno'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)


labels

adata1.obs['labels']=labels

pd.Series(adata1.obs['labels']).value_counts()

# Set the maximum sample size for downsampling
max_sample_size = 500

# Initialize a new DataFrame to store the downsampled data
downsampled_obs = pd.DataFrame()

# Iterate over unique categories and perform downsampling
for category in adata1.obs['labels'].unique():
    # Select data for the current category
    category_data = adata1[adata1.obs['labels'] == category]

    # Perform downsampling to the maximum sample size
    downsampled_category = resample(category_data.obs, replace=False, n_samples=min(len(category_data), max_sample_size), random_state=42)

    # Append the downsampled category to the new DataFrame
    downsampled_obs = pd.concat([downsampled_obs, downsampled_category])

# Create a new AnnData object with the downsampled observations
adata1 = adata1[downsampled_obs.index, :]

## Remove class that < 20 cells
# Identify categories that meet the threshold
label_counts = adata1.obs['labels'].value_counts()
valid_labels = label_counts[label_counts >= 20].index

# Filter AnnData based on the valid categories
adata1 = adata1[adata1.obs['labels'].isin(valid_labels)]

## Remove cell types
adata1 = adata1[adata1.obs['celltype.anno'] != "Primitive_erythroid"]
adata1 = adata1[adata1.obs['celltype.anno'] != "Testis_and_adrenal"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Definitive_erythroid"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Intermediate_neuronal_progenitors"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "White_blood_cells"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Adipocytes"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Eye_and_other"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Lung_and_airway"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Intestine"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "T_cells"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Olfactory_sensory_neurons"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "B_cells"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Mast_cells"]
#adata1 = adata1[adata1.obs['celltype.anno'] != "Oligodendrocytes"]

adata1

## Save the input data for GeneSys

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data1.pkl", 'wb') as file_handle:
    pickle.dump(adata1, file_handle)

del(adata1)

## adata2

adata2 = sc.read_h5ad("./Mouse_Embryo/adata_JAX_dataset_2.h5ad")

adata2

sc.pp.filter_cells(adata2, min_counts=2500)
sc.pp.filter_cells(adata2, min_genes=2500)
# Calculate row sums for the number of genes per cell
row_sums = np.sum(adata2.X > 0 , axis=1).reshape(-1).tolist()[0]
# Calculate row sums for the number of counts per cell
row_sums_counts = np.sum(adata2.X, axis=1).reshape(-1).tolist()[0]

np.median(row_sums)

np.median(row_sums_counts)

adata2

pd.Series(adata2.obs['day']).value_counts()

celltype = pd.merge(adata2.obs['cell_id'], df_cell, on='cell_id', how='left')
celltype.index = adata2.obs.index
adata2.obs['celltype.anno']=celltype['major_trajectory']

## Prepare labels
time_bin = adata2.obs['day'].copy()
time_bin = time_bin.replace('E8.5', 0)
time_bin = time_bin.replace('E8.75', 0)
time_bin = time_bin.replace('E9.0', 0)
time_bin = time_bin.replace('E9.25', 1)
time_bin = time_bin.replace('E9.5', 1)
time_bin = time_bin.replace('E9.75', 1)
time_bin = time_bin.replace('E10.0', 1)
time_bin = time_bin.replace('E10.25', 2)
time_bin = time_bin.replace('E10.5', 2)
time_bin = time_bin.replace('E10.75', 3)
time_bin = time_bin.replace('E11.0', 3)
time_bin = time_bin.replace('E11.25', 3)
time_bin = time_bin.replace('E11.5', 3)
time_bin = time_bin.replace('E11.75', 3)
time_bin = time_bin.replace('E12.0', 4)
time_bin = time_bin.replace('E12.25', 4)
time_bin = time_bin.replace('E12.5', 4)
time_bin = time_bin.replace('E12.75', 4)
time_bin = time_bin.replace('E13.0', 5)
time_bin = time_bin.replace('E13.25', 5)
time_bin = time_bin.replace('E13.5', 5)
time_bin = time_bin.replace('E13.75', 5)
time_bin = time_bin.replace('E14.0', 6)
time_bin = time_bin.replace('E14.25', 6)
time_bin = time_bin.replace('E14.333', 6)
time_bin = time_bin.replace('E14.75', 7)
time_bin = time_bin.replace('E15.0', 7)
time_bin = time_bin.replace('E15.25', 7)
time_bin = time_bin.replace('E15.5', 8)
time_bin = time_bin.replace('E15.75', 8)
time_bin = time_bin.replace('E16.0', 8)
time_bin = time_bin.replace('E16.25', 8)
time_bin = time_bin.replace('E16.5', 8)
time_bin = time_bin.replace('E16.75', 9)
time_bin = time_bin.replace('E17.0', 9)
time_bin = time_bin.replace('E17.25', 9)
time_bin = time_bin.replace('E17.5', 9)
time_bin = time_bin.replace('E17.75', 9)
time_bin = time_bin.replace('E18.0', 9)
time_bin = time_bin.replace('E18.25', 10)
time_bin = time_bin.replace('E18.5', 10)
time_bin = time_bin.replace('E18.75', 10)
time_bin = time_bin.replace('P0', 10)

celltype = adata2.obs['celltype.anno']
celltype = celltype.replace('Mesoderm', 0)
celltype = celltype.replace('CNS_neurons', 1)
celltype = celltype.replace('Neuroectoderm_and_glia', 2)
celltype = celltype.replace('Definitive_erythroid', 3)
celltype = celltype.replace('Epithelial_cells', 4)
celltype = celltype.replace('Intermediate_neuronal_progenitors', 5)
celltype = celltype.replace('Muscle_cells', 6)
celltype = celltype.replace('Endothelium', 7)
celltype = celltype.replace('White_blood_cells', 8)
celltype = celltype.replace('Hepatocytes', 9)
celltype = celltype.replace('Neural_crest_PNS_glia', 10)
celltype = celltype.replace('Neural_crest_PNS_neurons', 11)
celltype = celltype.replace('Adipocytes', 12)
celltype = celltype.replace('Eye_and_other', 13)
celltype = celltype.replace('Lung_and_airway', 14)
celltype = celltype.replace('Intestine', 15)
celltype = celltype.replace('T_cells', 16)
celltype = celltype.replace('Cardiomyocytes', 17)
celltype = celltype.replace('Olfactory_sensory_neurons', 18)
celltype = celltype.replace('B_cells', 19)
celltype = celltype.replace('Mast_cells', 20)
celltype = celltype.replace('Oligodendrocytes', 21)
celltype = celltype.replace('Megakaryocytes', 22)
celltype = celltype.replace('Ependymal_cells', 23)
celltype = celltype.replace('Primitive_erythroid', 24)
celltype = celltype.replace('Testis_and_adrenal', 25)                            
 
labels = adata2.obs['celltype.anno'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)

adata2.obs['labels']=labels

# Set the maximum sample size for downsampling
max_sample_size = 500

# Initialize a new DataFrame to store the downsampled data
downsampled_obs = pd.DataFrame()

# Iterate over unique categories and perform downsampling
for category in adata2.obs['labels'].unique():
    # Select data for the current category
    category_data = adata2[adata2.obs['labels'] == category]

    # Perform downsampling to the maximum sample size
    downsampled_category = resample(category_data.obs, replace=False, n_samples=min(len(category_data), max_sample_size), random_state=42)

    # Append the downsampled category to the new DataFrame
    downsampled_obs = pd.concat([downsampled_obs, downsampled_category])

# Create a new AnnData object with the downsampled observations
adata2 = adata2[downsampled_obs.index, :]

## Remove class that < 20 cells
# Identify categories that meet the threshold
label_counts = adata2.obs['labels'].value_counts()
valid_labels = label_counts[label_counts >= 20].index

# Filter AnnData based on the valid categories
adata2 = adata2[adata2.obs['labels'].isin(valid_labels)]

## Remove cell types
adata2 = adata2[adata2.obs['celltype.anno'] != "Primitive_erythroid"]
adata2 = adata2[adata2.obs['celltype.anno'] != "Testis_and_adrenal"]

pd.Series(adata2.obs['labels']).value_counts()

adata2

## Save the input data for GeneSys

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data2.pkl", 'wb') as file_handle:
    pickle.dump(adata2, file_handle)

del(adata2)

## adata3

adata3 = sc.read_h5ad("./Mouse_Embryo/adata_JAX_dataset_3.h5ad")

adata3

sc.pp.filter_cells(adata3, min_counts=2500)
sc.pp.filter_cells(adata3, min_genes=2500)
# Calculate row sums for the number of genes per cell
row_sums = np.sum(adata3.X > 0 , axis=1).reshape(-1).tolist()[0]
# Calculate row sums for the number of counts per cell
row_sums_counts = np.sum(adata3.X, axis=1).reshape(-1).tolist()[0]

np.median(row_sums)

np.median(row_sums_counts)

adata3

pd.Series(adata3.obs['day']).value_counts()

celltype = pd.merge(adata3.obs['cell_id'], df_cell, on='cell_id', how='left')
celltype.index = adata3.obs.index
adata3.obs['celltype.anno']=celltype['major_trajectory']

## Prepare labels
time_bin = adata3.obs['day'].copy()
time_bin = time_bin.replace('E8.5', 0)
time_bin = time_bin.replace('E8.75', 0)
time_bin = time_bin.replace('E9.0', 0)
time_bin = time_bin.replace('E9.25', 1)
time_bin = time_bin.replace('E9.5', 1)
time_bin = time_bin.replace('E9.75', 1)
time_bin = time_bin.replace('E10.0', 1)
time_bin = time_bin.replace('E10.25', 2)
time_bin = time_bin.replace('E10.5', 2)
time_bin = time_bin.replace('E10.75', 3)
time_bin = time_bin.replace('E11.0', 3)
time_bin = time_bin.replace('E11.25', 3)
time_bin = time_bin.replace('E11.5', 3)
time_bin = time_bin.replace('E11.75', 3)
time_bin = time_bin.replace('E12.0', 4)
time_bin = time_bin.replace('E12.25', 4)
time_bin = time_bin.replace('E12.5', 4)
time_bin = time_bin.replace('E12.75', 4)
time_bin = time_bin.replace('E13.0', 5)
time_bin = time_bin.replace('E13.25', 5)
time_bin = time_bin.replace('E13.5', 5)
time_bin = time_bin.replace('E13.75', 5)
time_bin = time_bin.replace('E14.0', 6)
time_bin = time_bin.replace('E14.25', 6)
time_bin = time_bin.replace('E14.333', 6)
time_bin = time_bin.replace('E14.75', 7)
time_bin = time_bin.replace('E15.0', 7)
time_bin = time_bin.replace('E15.25', 7)
time_bin = time_bin.replace('E15.5', 8)
time_bin = time_bin.replace('E15.75', 8)
time_bin = time_bin.replace('E16.0', 8)
time_bin = time_bin.replace('E16.25', 8)
time_bin = time_bin.replace('E16.5', 8)
time_bin = time_bin.replace('E16.75', 9)
time_bin = time_bin.replace('E17.0', 9)
time_bin = time_bin.replace('E17.25', 9)
time_bin = time_bin.replace('E17.5', 9)
time_bin = time_bin.replace('E17.75', 9)
time_bin = time_bin.replace('E18.0', 9)
time_bin = time_bin.replace('E18.25', 10)
time_bin = time_bin.replace('E18.5', 10)
time_bin = time_bin.replace('E18.75', 10)
time_bin = time_bin.replace('P0', 10)

celltype = adata3.obs['celltype.anno']
celltype = celltype.replace('Mesoderm', 0)
celltype = celltype.replace('CNS_neurons', 1)
celltype = celltype.replace('Neuroectoderm_and_glia', 2)
celltype = celltype.replace('Definitive_erythroid', 3)
celltype = celltype.replace('Epithelial_cells', 4)
celltype = celltype.replace('Intermediate_neuronal_progenitors', 5)
celltype = celltype.replace('Muscle_cells', 6)
celltype = celltype.replace('Endothelium', 7)
celltype = celltype.replace('White_blood_cells', 8)
celltype = celltype.replace('Hepatocytes', 9)
celltype = celltype.replace('Neural_crest_PNS_glia', 10)
celltype = celltype.replace('Neural_crest_PNS_neurons', 11)
celltype = celltype.replace('Adipocytes', 12)
celltype = celltype.replace('Eye_and_other', 13)
celltype = celltype.replace('Lung_and_airway', 14)
celltype = celltype.replace('Intestine', 15)
celltype = celltype.replace('T_cells', 16)
celltype = celltype.replace('Cardiomyocytes', 17)
celltype = celltype.replace('Olfactory_sensory_neurons', 18)
celltype = celltype.replace('B_cells', 19)
celltype = celltype.replace('Mast_cells', 20)
celltype = celltype.replace('Oligodendrocytes', 21)
celltype = celltype.replace('Megakaryocytes', 22)
celltype = celltype.replace('Ependymal_cells', 23)
celltype = celltype.replace('Primitive_erythroid', 24)
celltype = celltype.replace('Testis_and_adrenal', 25)                            
 
labels = adata3.obs['celltype.anno'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)

adata3.obs['labels']=labels

# Set the maximum sample size for downsampling
max_sample_size = 500

# Initialize a new DataFrame to store the downsampled data
downsampled_obs = pd.DataFrame()

# Iterate over unique categories and perform downsampling
for category in adata3.obs['labels'].unique():
    # Select data for the current category
    category_data = adata3[adata3.obs['labels'] == category]

    # Perform downsampling to the maximum sample size
    downsampled_category = resample(category_data.obs, replace=False, n_samples=min(len(category_data), max_sample_size), random_state=42)

    # Append the downsampled category to the new DataFrame
    downsampled_obs = pd.concat([downsampled_obs, downsampled_category])

# Create a new AnnData object with the downsampled observations
adata3 = adata3[downsampled_obs.index, :]

## Remove class that < 20 cells
# Identify categories that meet the threshold
label_counts = adata3.obs['labels'].value_counts()
valid_labels = label_counts[label_counts >= 20].index

# Filter AnnData based on the valid categories
adata3 = adata3[adata3.obs['labels'].isin(valid_labels)]

## Remove cell types
adata3 = adata3[adata3.obs['celltype.anno'] != "Primitive_erythroid"]
adata3 = adata3[adata3.obs['celltype.anno'] != "Testis_and_adrenal"]

adata3

## Save the input data for GeneSys

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data3.pkl", 'wb') as file_handle:
    pickle.dump(adata3, file_handle)

del(adata3)

## adata4

adata4 = sc.read_h5ad("./Mouse_Embryo/adata_JAX_dataset_4.h5ad")

adata4

sc.pp.filter_cells(adata4, min_counts=2500)
sc.pp.filter_cells(adata4, min_genes=2500)
# Calculate row sums for the number of genes per cell
row_sums = np.sum(adata4.X > 0 , axis=1).reshape(-1).tolist()[0]
# Calculate row sums for the number of counts per cell
row_sums_counts = np.sum(adata4.X, axis=1).reshape(-1).tolist()[0]

np.median(row_sums)

np.median(row_sums_counts)

adata4

pd.Series(adata4.obs['day']).value_counts()

celltype = pd.merge(adata4.obs['cell_id'], df_cell, on='cell_id', how='left')
celltype.index = adata4.obs.index
adata4.obs['celltype.anno']=celltype['major_trajectory']

## Prepare labels
time_bin = adata4.obs['day'].copy()
time_bin = time_bin.replace('E8.5', 0)
time_bin = time_bin.replace('E8.75', 0)
time_bin = time_bin.replace('E9.0', 0)
time_bin = time_bin.replace('E9.25', 1)
time_bin = time_bin.replace('E9.5', 1)
time_bin = time_bin.replace('E9.75', 1)
time_bin = time_bin.replace('E10.0', 1)
time_bin = time_bin.replace('E10.25', 2)
time_bin = time_bin.replace('E10.5', 2)
time_bin = time_bin.replace('E10.75', 3)
time_bin = time_bin.replace('E11.0', 3)
time_bin = time_bin.replace('E11.25', 3)
time_bin = time_bin.replace('E11.5', 3)
time_bin = time_bin.replace('E11.75', 3)
time_bin = time_bin.replace('E12.0', 4)
time_bin = time_bin.replace('E12.25', 4)
time_bin = time_bin.replace('E12.5', 4)
time_bin = time_bin.replace('E12.75', 4)
time_bin = time_bin.replace('E13.0', 5)
time_bin = time_bin.replace('E13.25', 5)
time_bin = time_bin.replace('E13.5', 5)
time_bin = time_bin.replace('E13.75', 5)
time_bin = time_bin.replace('E14.0', 6)
time_bin = time_bin.replace('E14.25', 6)
time_bin = time_bin.replace('E14.333', 6)
time_bin = time_bin.replace('E14.75', 7)
time_bin = time_bin.replace('E15.0', 7)
time_bin = time_bin.replace('E15.25', 7)
time_bin = time_bin.replace('E15.5', 8)
time_bin = time_bin.replace('E15.75', 8)
time_bin = time_bin.replace('E16.0', 8)
time_bin = time_bin.replace('E16.25', 8)
time_bin = time_bin.replace('E16.5', 8)
time_bin = time_bin.replace('E16.75', 9)
time_bin = time_bin.replace('E17.0', 9)
time_bin = time_bin.replace('E17.25', 9)
time_bin = time_bin.replace('E17.5', 9)
time_bin = time_bin.replace('E17.75', 9)
time_bin = time_bin.replace('E18.0', 9)
time_bin = time_bin.replace('E18.25', 10)
time_bin = time_bin.replace('E18.5', 10)
time_bin = time_bin.replace('E18.75', 10)
time_bin = time_bin.replace('P0', 10)

celltype = adata4.obs['celltype.anno']
celltype = celltype.replace('Mesoderm', 0)
celltype = celltype.replace('CNS_neurons', 1)
celltype = celltype.replace('Neuroectoderm_and_glia', 2)
celltype = celltype.replace('Definitive_erythroid', 3)
celltype = celltype.replace('Epithelial_cells', 4)
celltype = celltype.replace('Intermediate_neuronal_progenitors', 5)
celltype = celltype.replace('Muscle_cells', 6)
celltype = celltype.replace('Endothelium', 7)
celltype = celltype.replace('White_blood_cells', 8)
celltype = celltype.replace('Hepatocytes', 9)
celltype = celltype.replace('Neural_crest_PNS_glia', 10)
celltype = celltype.replace('Neural_crest_PNS_neurons', 11)
celltype = celltype.replace('Adipocytes', 12)
celltype = celltype.replace('Eye_and_other', 13)
celltype = celltype.replace('Lung_and_airway', 14)
celltype = celltype.replace('Intestine', 15)
celltype = celltype.replace('T_cells', 16)
celltype = celltype.replace('Cardiomyocytes', 17)
celltype = celltype.replace('Olfactory_sensory_neurons', 18)
celltype = celltype.replace('B_cells', 19)
celltype = celltype.replace('Mast_cells', 20)
celltype = celltype.replace('Oligodendrocytes', 21)
celltype = celltype.replace('Megakaryocytes', 22)
celltype = celltype.replace('Ependymal_cells', 23)
celltype = celltype.replace('Primitive_erythroid', 24)
celltype = celltype.replace('Testis_and_adrenal', 25)                            
 
labels = adata4.obs['celltype.anno'].astype(str) + "_" + time_bin.astype(str)
labels = celltype.astype(str) + "_" + time_bin.astype(str)

adata4.obs['labels']=labels

# Set the maximum sample size for downsampling
max_sample_size = 500

# Initialize a new DataFrame to store the downsampled data
downsampled_obs = pd.DataFrame()

# Iterate over unique categories and perform downsampling
for category in adata4.obs['labels'].unique():
    # Select data for the current category
    category_data = adata4[adata4.obs['labels'] == category]

    # Perform downsampling to the maximum sample size
    downsampled_category = resample(category_data.obs, replace=False, n_samples=min(len(category_data), max_sample_size), random_state=42)

    # Append the downsampled category to the new DataFrame
    downsampled_obs = pd.concat([downsampled_obs, downsampled_category])

# Create a new AnnData object with the downsampled observations
adata4 = adata4[downsampled_obs.index, :]

## Remove class that < 20 cells
# Identify categories that meet the threshold
label_counts = adata4.obs['labels'].value_counts()
valid_labels = label_counts[label_counts >= 20].index

# Filter AnnData based on the valid categories
adata4 = adata4[adata4.obs['labels'].isin(valid_labels)]

## Remove cell types
adata4 = adata4[adata4.obs['celltype.anno'] != "Primitive_erythroid"]
adata4 = adata4[adata4.obs['celltype.anno'] != "Testis_and_adrenal"]

adata4

## Save the input data for GeneSys

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data4.pkl", 'wb') as file_handle:
    pickle.dump(adata4, file_handle)

del(adata4)

## Combine data

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data1.pkl", 'rb') as file_handle:
    adata1 = pickle.load(file_handle)

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data2.pkl", 'rb') as file_handle:
    adata2 = pickle.load(file_handle)

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data3.pkl", 'rb') as file_handle:
    adata3 = pickle.load(file_handle)

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data4.pkl", 'rb') as file_handle:
    adata4 = pickle.load(file_handle)

df_gene = pd.read_csv('./Mouse_Embryo/df_gene.csv')

df_gene

adata1.var

adata2.var

adata3.var

adata4.var

adata = adata1.concatenate(adata2, adata3, adata4)

adata

del(adata1)
del(adata2)
del(adata3)
del(adata4)

## Processing all together
sc.pp.normalize_total(adata, target_sum=1e4)
sc.pp.log1p(adata)
sc.pp.scale(adata, max_value=10)
##The integrated assay values were floored at zero and normalized to the range of zero to one
##flooring
adata.X[adata.X < 0]=0
mmin = np.amin(adata.X)
nor = (np.amax(adata.X)-mmin)
adata.X = (adata.X-mmin)/nor

adata.X

## Prepare training, validation and test set
X_train, X_test, y_train, y_test = train_test_split(
            adata.X, adata.obs['labels'], test_size=0.2, stratify=adata.obs['labels'], shuffle=True, random_state=0)
X_train, X_val, y_train, y_val = train_test_split(
            X_train, y_train, test_size=0.2, stratify=y_train, shuffle=True, random_state=0)
data = {}
data['X_train'] = X_train
data['y_train'] = y_train
data['X_val'] = X_val
data['y_val'] = y_val
data['X_test'] = X_test
data['y_test'] = y_test

## Save the input data for GeneSys

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data.pkl", 'wb') as file_handle:
    pickle.dump(data, file_handle)

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data.pkl", 'rb') as file_handle:
    adata = pickle.load(file_handle)

## Check the input Data 

with open("./genesys_mouse_embryo_11_clusters_500_cells_per_state_2500_GPC_UPC_data.pkl", 'rb') as file_handle:
    adata = pickle.load(file_handle)

adata['X_train'].shape

adata['X_val'].shape

new_nonzero = adata['X_val'][~(adata['X_val']==0)]

len(adata['X_val'][~(adata['X_val']==0)]) / adata['X_val'].shape[0]*adata['X_val'].shape[1]


len(new_nonzero)/(adata['X_val'].shape[0]*adata['X_val'].shape[1])

adata['X_test'].shape

## Percentage of data used
round((64179+16045+20056)/11441407,2)

pd.Series(data['y_train']).value_counts()

full_time_bins = np.arange(11)

for i in np.arange(24):
    print(i)
    ct = sorted(pd.Series(np.array([i.split('_', 1)[1] for i in data['y_train']], dtype='int')[np.array([i.split('_', 1)[0] for i in data['y_train']], dtype='int')==i]).value_counts().index)
    print([x for x in full_time_bins if x not in ct])

pd.Series(data['y_val']).value_counts()

pd.Series(data['y_test']).value_counts()
