==> jupyter_notebook/11-GeneSys_cell_recovery_rate_estimation_20250728.ipynb <==
import os

## Set directory
os.chdir('/hpc/group/pbenfeylab/CheWei/CW_data/genesys')

import networkx as nx
from genesys_evaluate_v1 import *
%matplotlib inline
%config InlineBackend.figure_format = 'retina'

import seaborn as sns
import anndata

## Conda Env genesys on DCC
print(torch.__version__)
print(sc.__version__) 

## Genes considered/used (shared among samples) 
gene_list = pd.read_csv('./gene_list_1108.csv')

## Load data

with open("./genesys_root_data.pkl", 'rb') as file_handle:
    data = pickle.load(file_handle)
    
batch_size = 2000
dataset = Root_Dataset(data['X_test'], data['y_test'])
loader = DataLoader(dataset,
                         batch_size = batch_size,
                         shuffle = True, drop_last=True)
train_dataset = Root_Dataset(data['X_train'], data['y_train'])
train_loader = DataLoader(train_dataset,
                         batch_size = batch_size,
                         shuffle = True, drop_last=True)

with open("./genesys_rswt_data.pkl", 'rb') as file_handle:
    data = pickle.load(file_handle)

X_all = np.vstack((data['X_train'],data['X_val'],data['X_test']))
y_all = pd.concat((data['y_train'],data['y_val'],data['y_test']))
unseen_dataset = Root_Dataset_NoQC(X_all, y_all)
unseen_loader = DataLoader(unseen_dataset,
                         batch_size = batch_size,
                         shuffle = True, drop_last=True)

input_size = data['X_train'].shape[1]
## 10 cell types 
output_size = 10
embedding_dim = 256
hidden_dim = 256
n_layers = 2
device = "cpu"
path = "./"

## Load trained GeneSys model (Evaluate)

model = ClassifierLSTM(input_size, output_size, embedding_dim, hidden_dim, n_layers).to(device)
model.load_state_dict(torch.load(path+"/workstation/genesys_model_trained_on_root_atlas_20240308_continue4.pth", map_location=torch.device('cpu')))
model = model
model.eval()

### Sample data (2000 cells)

with open("./genesys_root_data.pkl", 'rb') as file_handle:
    data = pickle.load(file_handle)
    
batch_size = 2000
dataset = Root_Dataset(data['X_test'], data['y_test'])
loader = DataLoader(dataset,
                         batch_size = batch_size,
                         shuffle = True, drop_last=True)
train_dataset = Root_Dataset(data['X_train'], data['y_train'])
train_loader = DataLoader(train_dataset,
                         batch_size = batch_size,
                         shuffle = True, drop_last=True)

classes = ['Columella', 'Lateral Root Cap', 'Phloem', 'Xylem', 'Procambium', 'Pericycle', 'Endodermis', 'Cortex', 'Atrichoblast', 'Trichoblast']
class2num = {c: i for (i, c) in enumerate(classes)}
num2class = {i: c for (i, c) in enumerate(classes)}

sample = next(iter(loader))
xo = sample['x'].to(device)
y = sample['y'].to(device)
y_label = [num2class[i] for i in y.tolist()]

## 2000 cell type trajectories (11 dev stage) sampled, each stage has 17513 gene expression)
xo.shape

## How many cell type trajectories are sampled for each cell type?
pd.Series(y_label).value_counts()

### GEP impact on development

model = ClassifierLSTM(input_size, output_size, embedding_dim, hidden_dim, n_layers).to(device)
model.load_state_dict(torch.load(path+"/workstation/genesys_model_trained_on_root_atlas_20240308_continue4.pth", map_location=torch.device('cpu')))
model = model
model.eval()

GEP_raw_score = pd.read_csv('../../cNMF/20250722_root_atlas_cnmf_parallel.gene_spectra_score.k_30.dt_0_1.txt', delim_whitespace=True)
GEP_raw_score = GEP_raw_score.T

GEP_raw_score

# Find the GEP with the highest absolute loading for each gene
gep_assignments = GEP_raw_score.idxmax(axis=1)  # Get GEP with highest value
gep_scores = GEP_raw_score.lookup(GEP_raw_score.index, gep_assignments)  # Get actual values

# Add the assignment and score to the dataframe
GEP_score = pd.DataFrame({
    'Gene': GEP_raw_score.index,
    'GEP': gep_assignments,
    'Score': gep_scores
})

GEP_score

# Rank genes within each GEP by score
GEP_score['Rank'] = GEP_score.groupby('GEP')['Score'].rank(method='first', ascending=False).astype(int)

# Reset index and merge back the original data
GEP_score = GEP_score.merge(
    GEP_raw_score.reset_index(), left_on='Gene', right_on='index'
).drop(columns='index')

# Rename columns to match second image
gep_columns = [f'GEP_{i+1}' for i in range(GEP_raw_score.shape[1])]
GEP_score.columns = ['Gene', 'GEP', 'Score', 'Rank'] + gep_columns

GEP_score

# Remove rows with Score < or = 0
GEP_score = GEP_score[GEP_score['Score'] > 0]

GEP_score = GEP_score.sort_values(by=['GEP', 'Rank'], ascending=[True, True])
GEP_score['GEP'] = 'GEP_' + GEP_score['GEP'].astype(str)

GEP_score

GEP_score.to_csv('../../cNMF/20250722_root_atlas_cnmf_parallel.gene_spectra_score.k_30.dt_0_1_annotated.csv', index=False)

gene_list.columns = ['Gene'] + ['features']

## Keep genes that can only be found in GeneSys trained model
GEP_score = gene_list.merge(GEP_score, on='Gene', how='left')

GEP_score = GEP_score.sort_values(by=['GEP', 'Rank'], ascending=[True, True])

GEP_score

# Create a list of top genes for each GEP_ column
GEP_list = []
n_top = 50  # Number of top genes to select
# Loop through each GEP_ column (GEP_1 to GEP_31)
for i in range(1, 31):  # 1 to 30
    col = 'Rank'
    gep = f'GEP_{i}'
    # Sort by the current GEP_ column in descending order, select top 20, and get gene_IDs
    top_genes = GEP_score[GEP_score['GEP'] == gep].sort_values(by=col, ascending=True).head(n_top)['Gene'].tolist()
    GEP_list.append(top_genes)
# Name the list elements (equivalent to names(GEP_list) in R)
GEP_dict = {f'GEP_{i}': genes for i, genes in enumerate(GEP_list, 1)}

{key: len(value) for key, value in GEP_dict.items()}

## Define functions

def recovery(matched_idx, output_dir_and_file_name):
    y_pred, y_true = [], []
    
    with torch.no_grad():
        for sample in loader:
            x = sample['x'].to(device)
            
            # remove/keep genes
            to_keep = x[:,:,matched_idx]
            # remove all gene expression
            x[:,:,:]=x[:,:,:].zero_()
            # add back the gene expression for selected genes
            x[:,:,matched_idx]= to_keep
            
            x10 = torch.stack([x[:,0,:],x[:,1,:],x[:,2,:],x[:,3,:],x[:,4,:],x[:,5,:],x[:,6,:],x[:,7,:],x[:,8,:],x[:,9,:],x[:,10,:]],dim=1)
            y = sample['y'].to(device)
            y_true.append(y.cpu().detach().numpy())
            test_h = model.init_hidden(batch_size)
            p, pred_h = model.predict_proba(x10, test_h, 10)
            y_pred.append(p.cpu().detach().numpy())
    
            
    y_true = np.concatenate(y_true)
    y_pred = np.concatenate(y_pred)

    # Compute overall recovery
    overall_recovery = (y_true == np.argmax(y_pred, axis=1)).mean()
    
    # Store results in a dictionary
    results = {'Celltype': ['Overall'], 'Recovery': [overall_recovery]}
    
    # Define cell types
    classes = ['Columella', 'Lateral Root Cap', 'Phloem', 'Xylem', 'Procambium',
               'Pericycle', 'Endodermis', 'Cortex', 'Atrichoblast', 'Trichoblast']
    
    # Compute recovery per cell type
    for ct in range(10):
        idx = np.where(y_true == ct)
        recovery = (y_true[idx] == np.argmax(y_pred, axis=1)[idx]).mean()
        results['Celltype'].append(classes[ct])
        results['Recovery'].append(recovery)
    
    # Convert to DataFrame
    df = pd.DataFrame(results)
    print(df)
    
    # Save to CSV
    df.to_csv(output_dir_and_file_name, index=False)

def t_recovery(matched_idx, time):
    y_pred, y_true = [], []
    
    with torch.no_grad():
        for sample in loader:
            x = sample['x'].to(device)
            
            # remove/keep genes
            to_keep = x[:,:,matched_idx]
            # remove all gene expression
            x[:,:,:]=x[:,:,:].zero_()
            # add back the gene expression for selected genes
            x[:,:,matched_idx]= to_keep
            
            x10 = torch.stack([x[:,0,:],x[:,1,:],x[:,2,:],x[:,3,:],x[:,4,:],x[:,5,:],x[:,6,:],x[:,7,:],x[:,8,:],x[:,9,:],x[:,10,:]],dim=1)
            y = sample['y'].to(device)
            y_true.append(y.cpu().detach().numpy())
            test_h = model.init_hidden(batch_size)
            p, pred_h = model.predict_proba(x10, test_h, time)
            y_pred.append(p.cpu().detach().numpy())
    
            
    y_true = np.concatenate(y_true)
    y_pred = np.concatenate(y_pred)

    # Compute overall recovery
    overall_recovery = (y_true == np.argmax(y_pred, axis=1)).mean()
    
    # Store results in a dictionary
    results = {'Celltype': ['Overall'], 'Recovery_t' + str(int(time)): [overall_recovery]}
    
    # Define cell types
    classes = ['Columella', 'Lateral Root Cap', 'Phloem', 'Xylem', 'Procambium',
               'Pericycle', 'Endodermis', 'Cortex', 'Atrichoblast', 'Trichoblast']
    
    # Compute recovery per cell type
    for ct in range(10):
        idx = np.where(y_true == ct)
        recovery = (y_true[idx] == np.argmax(y_pred, axis=1)[idx]).mean()
        results['Celltype'].append(classes[ct])
        results['Recovery_t' + str(int(time))].append(recovery)
    
    # Convert to DataFrame
    df = pd.DataFrame(results)
    #print(df)
    
    # Return df
    return df

def gof(matched_idx):
    # Find indices where query_list elements match
    #matched_idx = [i for i, x in enumerate(gene_list['features']) if x in set(GEP_dict[GEP])]
    matched_idx = matched_idx
    #len(gene_list['features'][matched_idx])
    
    #Prepare 
    xm = xo.clone()
    # remove/keep genes
    to_keep = xm[:,:,matched_idx]
    xm[:,:,:]=xm[:,:,:].zero_()
    xm[:,:,matched_idx]= to_keep
    ## Provide entire tracks
    x = torch.stack([xm[:,0,:],xm[:,1,:],xm[:,2,:],xm[:,3,:],xm[:,4,:],xm[:,5,:],xm[:,6,:],xm[:,7,:],xm[:,8,:],xm[:,9,:],xm[:,10,:]],dim=1)
    ## Provide the first two bins 
    #x = torch.stack([xm[:,0,:],xm[:,1,:],xm[:,2,:],xm[:,2,:],xm[:,2,:],xm[:,2,:],xm[:,2,:],xm[:,2,:],xm[:,2,:],xm[:,2,:],xm[:,2,:]],dim=1)

    ## Initialize hidden state
    pred_h = model.init_hidden(batch_size)
    
    # t0 and t1 prediction based on data from the first two time points
    t0 = model.generate_current(x, pred_h, 0)
    
    ## predict t1 label
    y0, pred_h = model.predict_proba(x, pred_h, 0)
    y0 = [num2class[i] for i in np.argmax(y0.cpu().detach().numpy(), axis=1)]
    
    t1 = model.generate_next(x, pred_h, 0)
    
    y1, pred_h = model.predict_proba(x, pred_h, 1)
    y1 = [num2class[i] for i in np.argmax(y1.cpu().detach().numpy(), axis=1)]

    t2 = model.generate_next(x, pred_h, 1)
    
    y2, pred_h = model.predict_proba(x, pred_h, 2)
    y2 = [num2class[i] for i in np.argmax(y2.cpu().detach().numpy(), axis=1)]
    
    t3 = model.generate_next(x, pred_h, 2)
    
    y3, pred_h = model.predict_proba(x, pred_h, 3)
    y3 = [num2class[i] for i in np.argmax(y3.cpu().detach().numpy(), axis=1)]
    
    t4 = model.generate_next(x, pred_h, 3)
    
    y4, pred_h = model.predict_proba(x, pred_h, 4)
    y4 = [num2class[i] for i in np.argmax(y4.cpu().detach().numpy(), axis=1)]

    t5 = model.generate_next(x, pred_h, 4)
    
    y5, pred_h = model.predict_proba(x, pred_h, 5)
    y5 = [num2class[i] for i in np.argmax(y5.cpu().detach().numpy(), axis=1)]
    
    t6 = model.generate_next(x, pred_h, 5)
    
    y6, pred_h = model.predict_proba(x, pred_h, 6)
    y6 = [num2class[i] for i in np.argmax(y6.cpu().detach().numpy(), axis=1)]
    
    t7 = model.generate_next(x, pred_h, 6)

    y7, pred_h = model.predict_proba(x, pred_h, 7)
    y7 = [num2class[i] for i in np.argmax(y7.cpu().detach().numpy(), axis=1)]
    
    t8 = model.generate_next(x, pred_h, 7)
    
    y8, pred_h = model.predict_proba(x, pred_h, 8)
    y8 = [num2class[i] for i in np.argmax(y8.cpu().detach().numpy(), axis=1)]
    
    t9 = model.generate_next(x, pred_h, 8)

    y9, pred_h = model.predict_proba(x, pred_h, 9)
    y9 = [num2class[i] for i in np.argmax(y9.cpu().detach().numpy(), axis=1)]
    
    t10 = model.generate_next(x, pred_h, 9)
    
    y10, pred_h = model.predict_proba(x, pred_h, 10)
    y10 = [num2class[i] for i in np.argmax(y10.cpu().detach().numpy(), axis=1)]

    t0 = t0.to(device).detach().numpy()
    t1 = t1.to(device).detach().numpy()
    t2 = t2.to(device).detach().numpy()
    t3 = t3.to(device).detach().numpy()
    t4 = t4.to(device).detach().numpy()
    t5 = t5.to(device).detach().numpy()
    t6 = t6.to(device).detach().numpy()
    t7 = t7.to(device).detach().numpy()
    t8 = t8.to(device).detach().numpy()
    t9 = t9.to(device).detach().numpy()
    t10 = t10.to(device).detach().numpy()
    
    pred_X = np.vstack((t0,t1,t2,t3,t4,t5,t6,t7,t8,t9,t10))
    pred_Y = np.concatenate((['Stem Cell']*batch_size, y1, y2, y3, y4, y5, y6, y7, y8, y9, y10)).tolist()
    pred_T = ['t0']*batch_size + ['t1']*batch_size + ['t2']*batch_size + ['t3']*batch_size + ['t4']*batch_size + ['t5']*batch_size + ['t6']*batch_size + ['t7']*batch_size + ['t8']*batch_size+ ['t9']*batch_size + ['t10']*batch_size

    # Create AnnData object
    cell_names = [f"Cell_{i}" for i in range(pred_X.shape[0])]
    adata = anndata.AnnData(
        X=pred_X,
        obs=pd.DataFrame(index=cell_names),  # Cell annotations
        var=pd.DataFrame(index=gene_list['features'])   # Gene annotations
    )
    adata.obs['celltype'] = pred_Y
    adata.obs['timebin'] = pred_T

    sc.pp.scale(adata, max_value=10)
    sc.tl.pca(adata, svd_solver='arpack')
    sc.pp.neighbors(adata, n_neighbors=30, n_pcs=50)
    sc.tl.leiden(adata)
    sc.tl.paga(adata)
    sc.pl.paga(adata) 

    sc.tl.umap(adata, init_pos='paga')
    adata.uns['celltype_colors'] = np.array([ "#9400d3","#5ab953", "#bfef45", "#008080", "#21B6A8", "#82b6ff", "#0000FF","#e6194b", "#9a6324", "#ffe119", "#ff9900", "#ffd4e3", "#9a6324", "#ddaa6f"], dtype=object)
    adata.obs['celltype'] = pd.Categorical(adata.obs['celltype'], categories=["Stem Cell","Columella", "Lateral Root Cap", "Atrichoblast", "Trichoblast", "Cortex", "Endodermis", "Phloem", "Xylem", "Procambium","Pericycle"])
    sc.pl.umap(adata, color=['celltype'])

    adata.uns['timebin_colors'] = np.array([ '#5E4FA2', '#3288BD', '#66C2A5', '#ABDDA4', '#E6F598', '#FFFFBF', '#FEE08B', '#FDAE61', '#F46D43', '#D53E4F','#9E0142'])
    sc.pl.umap(adata, color=['timebin'])

    ## AT1G79840
    sc.pl.umap(adata, color='AT1G79840', title='AT1G79840 (GL2, Atrichoblast)')
    ## AT5G49270
    sc.pl.umap(adata, color='AT5G49270', title='AT5G49270 (COBL9, Trichoblast)')
    sc.pl.umap(adata, color='AT1G09750', title='AT1G09750 (CORTEX, Cortex)')
    sc.pl.umap(adata, color='AT5G57620', title='AT5G57620 (MYB36, Endodermis)')
    sc.pl.umap(adata, color='AT1G79430', title='AT1G79430 (APL, Phloem)')
    sc.pl.umap(adata, color='AT1G71930', title='AT1G71930 (VND7, Xylem)')
    #sc.pl.umap(adata, color='AT4G37650', title='AT4G37650 (SHR)')    

    return adata

## TFs

TF = pd.read_csv("../../CW_data/celloracle/Kay_TF_thalemine_annotations.csv")

TF.columns = ['Gene'] + ['Name'] + ['Description']

TF

## Top 30

# Create a list of top genes for each GEP_ column
GEP_list = []
n_top = 30  # Number of top genes to select
# Loop through each GEP_ column (GEP_1 to GEP_31)
for i in range(1, 31):  # 1 to 30
    col = 'Rank'
    gep = f'GEP_{i}'
    # Sort by the current GEP_ column in descending order, select top 20, and get gene_IDs
    top_genes = GEP_score[GEP_score['GEP'] == gep].sort_values(by=col, ascending=True).head(n_top)['Gene'].tolist()
    GEP_list.append(top_genes)
# Name the list elements (equivalent to names(GEP_list) in R)
GEP_dict = {f'GEP_{i}': genes for i, genes in enumerate(GEP_list, 1)}

##### GEP recovery

for g in GEP_dict.keys():
    matched_idx = [i for i, x in enumerate(gene_list['features']) if x in set(GEP_dict[g])]
    print(len(matched_idx))
    recovery(matched_idx, "./" + g + "_celltype_recovery.csv")

import glob
# Locate all recovery CSV files
file_list = sorted(glob.glob("./GEP_*_celltype_recovery.csv"))

# Read and merge
merged_df = None

for file in file_list:
    # Extract GEP name from filename (e.g., 'GEP_1')
    gep_name = os.path.basename(file).split('_celltype')[0]
    
    # Read the CSV
    df = pd.read_csv(file)
    
    # Rename "Recovery" column to current GEP
    df = df.rename(columns={'Recovery': gep_name})
    
    if merged_df is None:
        merged_df = df
    else:
        merged_df = pd.merge(merged_df, df, on='Celltype', how='outer')

# Optional: Set Celltype as index
merged_df = merged_df.set_index('Celltype')

# Sort columns by GEP number if desired
merged_df = merged_df[sorted(merged_df.columns, key=lambda x: int(x.split('_')[1]))]

# Define your custom order
custom_order = [
    "Xylem", "Phloem", "Procambium", "Pericycle", "Endodermis",
    "Cortex", "Trichoblast", "Atrichoblast", "Lateral Root Cap",
    "Columella", "Overall"
]

# Make sure index is Celltype
merged_df.index = pd.Categorical(merged_df.index, categories=custom_order, ordered=True)

# Sort by the categorical index
merged_df = merged_df.sort_index()

merged_df

merged_df.to_csv('./Merged_GEP_celltype_recovery.csv', index=True)

## Test GEP 17 and 20 (top30)

# Without GEP17 and GEP20 
matched_idx = [i for i, x in enumerate(gene_list['features']) if x in set(GEP_dict['GEP_1'] + GEP_dict['GEP_3'] + GEP_dict['GEP_4'] 
                                                                          + GEP_dict['GEP_6'] + GEP_dict['GEP_8'] + GEP_dict['GEP_10'] 
                                                                          + GEP_dict['GEP_11'] + GEP_dict['GEP_12']+ GEP_dict['GEP_13']
                                                                          + GEP_dict['GEP_16']
                                                                          + GEP_dict['GEP_19'] + GEP_dict['GEP_22']+ GEP_dict['GEP_23']
                                                                          + GEP_dict['GEP_26'] + GEP_dict['GEP_27'])]

len(matched_idx)

## Without 17 and 20
GEP_com_wo_17_20 = recovery(matched_idx, "./GEP_com_wo_17_20celltype_recovery.csv")

# With 17
matched_idx = [i for i, x in enumerate(gene_list['features']) if x in set(GEP_dict['GEP_1'] + GEP_dict['GEP_3'] + GEP_dict['GEP_4'] 
                                                                          + GEP_dict['GEP_6'] + GEP_dict['GEP_8'] + GEP_dict['GEP_10'] 
                                                                          + GEP_dict['GEP_11'] + GEP_dict['GEP_12']+ GEP_dict['GEP_13']
                                                                          + GEP_dict['GEP_16'] + GEP_dict['GEP_17']
                                                                          + GEP_dict['GEP_19'] + GEP_dict['GEP_22']+ GEP_dict['GEP_23']
                                                                          + GEP_dict['GEP_26'] + GEP_dict['GEP_27'])]

len(matched_idx)

## With 17
GEP_com_with_17 = recovery(matched_idx, "./GEP_com_with_17_celltype_recovery.csv")

# With 20
matched_idx = [i for i, x in enumerate(gene_list['features']) if x in set(GEP_dict['GEP_1'] + GEP_dict['GEP_3'] + GEP_dict['GEP_4'] 
                                                                          + GEP_dict['GEP_6'] + GEP_dict['GEP_8'] + GEP_dict['GEP_10'] 
                                                                          + GEP_dict['GEP_11'] + GEP_dict['GEP_12']+ GEP_dict['GEP_13']
                                                                          + GEP_dict['GEP_16'] + GEP_dict['GEP_20']
                                                                          + GEP_dict['GEP_19'] + GEP_dict['GEP_22']+ GEP_dict['GEP_23']
                                                                          + GEP_dict['GEP_26'] + GEP_dict['GEP_27'])]

len(matched_idx)

## With 20
GEP_com_with_20 = recovery(matched_idx, "./GEP_com_with_20_celltype_recovery.csv")
